Refactor chat view and enhance user store
Updated the ChatView.vue to use Vue's computed property instead of the ref and watch combination. This makes the code cleaner and easier to understand. Also, the user store was improved by expanding the user and conversation data structure, including the addition of user id and user names. Moreover, visual modifications were made to the chat side menu, such as color tweaks and addition of the user's name.