Add scrolling to replied messages feature
The commit includes the addition to the message list component of a function to scroll to a replied message, identified by id. This is triggered when the reply reference in a message is clicked. Code in the MessageListItem component has been refactored for a cleaner implementation, providing a more streamlined and smoother user experience.