Add identifier to chat input and implement auto-focus
The changes in this commit include the addition of an 'id' attribute to the chat text input box to facilitate identification. Subsequently, an auto-focus feature is implemented in the 'ReplyToMessage' and 'EditMessage' functions. When these functions are invoked, the chat input box will automatically gain focus, thereby improving user experience.