Refactor mock message generation and retrieve older messages
The mock message generation function has been updated for improved flexibility and consistency. It now accepts the amount of messages to generate as a parameter and includes a time offset. The user store now contains a function to get older messages. The loading of older messages has been moved to the MessageList component, ensuing messages are loaded with the component.