Add newer message handling and visible message store
An additional function `GetNewerMessages` has been added to the `user.js` store to allow retrieval of new messages from a certain point. A new `messageList.js` store has been created to keep track of visible messages in a conversation. Changes in the `MessageList.vue` have been made to accommodate this new method of handling messages. These include loading and sorting of messages and observing mutations to update scroll position as messages are added.