Add Side Menu Toggle Button & Search Store
Added a left side menu toggle button to the chat view and friend list view. Along with this, a new Vuex store named "search" has been created to hold the search data and status. This update also includes some enhancements to the chat view to add search form and to update UI based on search status. The action of toggling the side menu previously in the "ChatTopBar" has been moved to a new component "ChatLeftSideMenuButton".